Rename markerCache.ts to tileMarkerCache.ts
