+++ overwolf/src/Minimap.tsx
@@ -5,9 +5,9 @@ import { globalLayers } from './globalLayers';
 import { registerEventCallback } from './logic/hooks';
 import { getIcon, GetPlayerIcon, setIconScale } from './logic/icons';
 import { getMapTiles } from './logic/map';
-import { getTileMarkerCache } from './logic/markerCache';
 import { getMarkers } from './logic/markers';
 import { getTileCache } from './logic/tileCache';
+import { getTileMarkerCache } from './logic/tileMarkerCache';
 import { toMinimapCoordinate } from './logic/tiles';
 import { rotateAround } from './logic/util';
 import { makeStyles } from './theme';
